% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastaccuracy.R
\name{forecastaccuracy}
\alias{forecastaccuracy}
\title{Input predictions and observations}
\usage{
forecastaccuracy(pred, obs)
}
\arguments{
\item{infile}{Predictions and Observations}
}
\value{
Forecasting Model Accuarcy in percentage
}
\description{
A function to calculate accuracy of timeseries forecasting models
Created by Sulekha Aloorravi
}
